pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-presentation-toolkit"
version = "0.1.0"
description = "A generic toolkit for migrating PowerPoint presentations to branded templates"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Office Suites",
]
keywords = ["powerpoint", "pptx", "presentation", "migration", "branding"]

dependencies = [
    "lxml>=4.9.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
pdf = ["PyMuPDF>=1.23.0"]
analyze = ["python-pptx>=0.6.21"]
all = [
    "PyMuPDF>=1.23.0",
    "python-pptx>=0.6.21",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
pptx-migrate = "presentation_toolkit.cli:pptx_migrate"
pptx-analyze = "presentation_toolkit.cli:pptx_analyze"
pptx-extract = "presentation_toolkit.cli:pptx_extract"
presentation-toolkit = "presentation_toolkit.cli:main"

[project.urls]
Homepage = "https://github.com/zivtech/claude-presentation-toolkit"
Repository = "https://github.com/zivtech/claude-presentation-toolkit"
Issues = "https://github.com/zivtech/claude-presentation-toolkit/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]  # Line length handled by black
